redux is hooked in
